Mac OS 10.5 compile fixes.

+++ examples/Sc21FullscreenViewer/AppController.mm
@@ -78,8 +78,8 @@
 {
   CGLContextObj cglContext;
   CGDisplayErr err;
-  long oldSwapInterval;
-  long newSwapInterval;
+  GLint oldSwapInterval;
+  GLint newSwapInterval;
 
   SCOpenGLPixelFormat * newformat = [[[SCOpenGLPixelFormat alloc] init] autorelease];
   [newformat setAttribute:NSOpenGLPFAAccelerated];
